Refactor reference to 'children'
As 'BottomSheetConsumer' was removed in https://github.com/WordPress/gutenberg/pull/33845/commits/523239866d5b502672695af6daa30453b57d26ac, 'children' no longer needs to be returned as part of a function. In fact, this results in an error.

With this commit, the function surrounding the reference to 'children' is removed.