Add title style to "Featured Image" panel
As we need to tweak the spacing around the panel's title, I've added a new "titleStyle" prop to the PanelBody component.

That prop is then used to remove the bottom spacing for the "Featured Image" panel's title.

The prop can now be re-used by anyone else who needs to tweak the title in a PanelBody in the future.