Latest Comments in-editor preview of comment excerpts match front-end (#33944)
When comment excerpts are included in the Latest Comments block, the
truncation is done differently depending on the site's locale (see
`get_comment_excerpt()` and `wp_trim_words()`). However when rendering
the block preview in the editor the preview is generated using the
user's locale. These rules can result in a different preview on the
front-end and the editor e.g. if the editing user's locale is `en` and
the site's locale is `zh-tw`.

This change ensures the in-editor preview of Latest Comments uses the
site's locale, just as it would appear on the front-end. Specifying an
empty `_locale` param bypasses the user-locale middleware which usually
sets the param to `_locale=user`.

The endpoint doesn't actually interpret the `'site'` value as anything
special, it's just a way of self-documenting what the code is trying to
achieve. The endpoint interprets it as an unrecognised locale slug and
falls back to the default (which is the site's locale).