Button: update spacing support to use axial padding (#33859)
* Add support for axial sides to padding block support

* Update button to support general and axial padding only

* Update docs

* Enable axial support for margin block support

* Warn against invalid sides configuration and disable support

If a spacing support is configured with an invalid mix of axial and
individual sides, log a warning and disable the feature. If only
one feature is invalidly configured, this does not affect the other;
eg if padding is configured incorrectly but margin is not, padding
is disabled but margins will continue to work.

* Update docs

* Clearer hook name

* Check that support is defined in useCustomSides

Previously the hook was only used if the feature was known to be
supported. Now that it is used in checking the spacing configuration,
it is possible for `support` to be undefined here.

* Update global styles panel

* Make padding a default control for the button block