Help section displays a full screen bottom sheet
Improve readability of large quantities of help content by leveraging a
full screen height bottom sheet.

In order to render the nested `ScrollView` at the correct height,
`flexShrink: 1` was added to the bottom sheet header element and the
`listProps`. The latter is applied to the bottom sheet child scrolling
element.

Additionally, `height: 100%` must be applied to the top-level element of
the child view rendered in this bottom sheet. This matches the link
picker implementation and ensures the element fills the parent bottom
sheet.