Editor Onboarding: "The Basics" help section - Details Part 2 (#34018)
* Refactor Editor help to use BottomSheet.SubSheet component

Avoid duplicative code by leveraging the existing abstraction.

* Remove unused image asset

* Improve accessibility of headings

Adding the "header" role provides improved audible cues of the content
structure.

* Improve accessibility of visual illustration

This particular image attempts to communicate the location of the text
formatting controls.

* Avoid cursor input language within help copy

Clicking is an action specific to cursor input devices, which is not
relevant for touch input devices.

* Improve screen reader navigation within bottom sheet

Declaring text elements as headers where appropriate improves the
ability to navigate and understand content structure.

* Update styles to align with designs

Various spacing and sizing changes were required to match the design.

* Affix Help Section heading to top of bottom sheet

To improve navigation, scroll the help content separately from the
heading that remains fixed atop the bottom sheet.

* Update image assets

Fix a few broken images. Include assets for higher pixel density
displays.

* Improve dark mode styles

* Update title font weight to match design

This matches the design comp and the title found elsewhere in other
bottom sheets.

* Help section displays a full screen bottom sheet

Improve readability of large quantities of help content by leveraging a
full screen height bottom sheet.

In order to render the nested `ScrollView` at the correct height,
`flexShrink: 1` was added to the bottom sheet header element and the
`listProps`. The latter is applied to the bottom sheet child scrolling
element.

Additionally, `height: 100%` must be applied to the top-level element of
the child view rendered in this bottom sheet. This matches the link
picker implementation and ensures the element fills the parent bottom
sheet.

* Replace BottomSheet.SubSheet with individual React Navigation screens

The current `BottomSheet.SubSheet` implementation does not reset a given
sub sheets visibility when navigating backwards via the Android back
button. This results in the child content remaining visible. When
subsequently navigating to a new child, the previous child is also
displayed.

We might be able to hook into React Navigation events to resolve this,
but my exploration felt forced and not all that valuable.

* Add close button for Help section

Including a close button will make it easier to dismiss the Help section
on Android.

* Enable default platform navigation transitions and gesture

Better align with OS defaults to meet user expectations.

* Refactor to avoid unnecessary code and changes

Some code is unused, other changes were not necessary or discussed with
the original author.

* Help section title adapts to editor context

Replace "site" with "post" or "page" base on the current content type.
We may reinstate the term "site" once full-site editing features are
included in the help articles.

* Remove "The basics" panel title

The panel title provides more confusion than value as we only have one
panel currently. Once we have multiple panels, reinstating the title
would be helpful.

* Enable text selection for help content

Allow selecting and copying text from help content to improve ability to
copy terms into search engines and elsewhere.

* Increase size of NavigationHeaader tap targets

Improve usability of UI via larger tap targets for buttons.