Fix Animated warning log
Animated now requires setting an explicit `useNativeDriver` option.
Without setting it, the following warning is produced.

```
WARN  Animated: `useNativeDriver` was not specified. This is a required option and must be explicitly set to `true` or `false`
```

It cannot be set to `true` for this particular animation as it is not
supported and produces the following error.

```
ERROR  Error: Style property 'left' is not supported by native animated module
```