Link picker dismisses iOS keyboard to ensure smooth animation
Focus of the URL text input caused animation stutter while the iOS
keyboard visibility toggled from hide to show to hide. To address this,
the Android-specific keyboard dismiss logic is now applied to iOS as
well.

When the keyboard visibility changes, the `KeyboardingAvoidingView`
configures a `LayoutAnimation` to manage the change. The quick toggle of
hide to show to hide could result in an unconsumed `LayoutAnimation`
lingering as no changes occurred which would consume the
`LayoutAnimation`.

Unconsumed `LayoutAnimations` appear to cause issues for React Native's
Modal component on iOS. Specifically, it can result in a transparent,
non-dismissible modal sitting atop the rest of the app UI. Upgrading to
React Native 0.66 appears to have increased the frequency of this
occurring, for an unknown reason.

- https://git.io/J1W3z
- https://git.io/J1W32
- https://git.io/J1W3a
- https://git.io/J1W3r
- https://git.io/J1W36