Fix missing animations for full-screen bottom sheets on Android
The `layout` conditional was removed in https://git.io/JEDdk, as it is
no longer utilized. However, its prior presence was universally
disabling the block it guarded. Removing the always falsey conditional
caused the block to always run and break the Android animation.

Specifically, when navigating from a full-screen bottom sheet to a
non-full-screen bottom sheet, the `currentHeight` is always `1`. This
results in setting the height without layout animations prior to this
change.