Fix React Native warning caused by webpack-oriented Jest image mock
The previous image mock was implemented following the Jest documentation
related to webpack. However, the `source` prop for React Native's Image
component does not accept strings. The previous image mock resulted in
the following warning in testing logs.

```
Warning: Failed prop type: Invalid prop `source` supplied to `Image`.
```

To address this, the previous image mock was removed in favor of relying
upon the transform that is provided by the `react-native` Jest preset.
In order to have this transform applied to images within React
Navigation, additional regex was added to `transformIgnorePatterns`.

- https://git.io/JEFo0
- https://bit.ly/3jAqaJL
- https://bit.ly/2YiuDbS
- https://git.io/JEFor
- https://git.io/JEFoK