Use symbol icon for reusable block in all cases
In case the reusable block only contains a single block, the icon of that block is used when presenting the reusable block in the block picker. This change limits this behaviour to the web version of the editor, so now in the native version, it will always use the symbol icon.