Adapt ColorPanel hook for native use
This commit takes the web's ColorPanel hook and adapts it for native use.

Specifically, while the web version relies on ComputedStyles to fetch a block's background/text colour, we need to use other means for native.

Here, we're using the useGlobalStyles, which will fetch a block's background and text colour when they're explicitly set via the global style settings.