Verify npm cache before a clean install (#34973)
The `Build Release Artifact` step of the `Build Plugin Zip` GH action sometimes fails because of `npm` cache issues. Namely, it fails with a `npm ERR cb() never called` error message. This attempt to verify the cache before the `ci` to give `npm` a chance to repair it, if needed.