Do not alter attributes of block being duplicated
`__experimentalRemoveAttributesByRole` takes a block instance and removes
any attributes with the given `role`. When cloning a block, we want to
remove internal attributes on the clone, but not the original copy. This
changes the order of operations in the cloning function to make sure that
those attributes only get removed on the latter.

Without this change, when a block is duplicated in the widget editor all
internal attributes will be stripped from BOTH the copy and the duplicate.