Use `__experimentalCloneSanitizedBlock` in copy handlers
When a block is duplicated, we remove internal attributes and then fill in
any default values (when we call `__experimentalSanitizeBlockAttributes`).
When we copy a block, we were only doing the first step of removing the
internal attrs, but not filling the defaults.

This causes a block validation issue, because the serialized block in the
clipboard is missing internal attrs which will get filled in by `createBlock`
when we actually paste.

This commit updates the copy handlers to use the cloneBlock code so that the
behavior matches block duplication exactly. It also reverts the
`__experimentalRemoveAttributesByRole` to the original API that operates on
an attributes object rather than a block instance, to match the signature of
`__experimentalSanitizeBlockAttributes` and clean up the logic in the
cloning function.