Hide block's caption component in Tiled Gallery
As the image block's caption component isn't part of the Tiled Gallery's design, this commit sets up the necessary code to hide it from view.

This is done through the use of a "hideImageCaption" context, which lives under a new "blockStyles" object. It should be possible to further utilise this object for future styling of inner blocks.