Only used fixed height if it's explicitly set
At the moment, a fixed height is applied to the image block if context is detected. This not only effects the standard Gallery block (intended) but also the Tiled Gallery Block (not intended).

With this commit, a new "fixedHeight" context is added to only set a fixed height if it's explicitly set.

This is necessary as the Tiled Gallery block will need varying heights as part of its design.