Allow third-party blocks to set "shapeStyle"
In order to change the image's shape via a third-party block (such as Jetpack's Tiled Gallery), we need to be able to pass a style to the "shapeStyle" prop.

With this commit, the "shapeStyle" prop is changed to accept both a reference to the image block's classes and a third-party block's styles.

This change means that it's no longer necessary for third-party blocks, like the Tiled Gallery, to set styles directly in the image block's stylesheet. So, this commit also removes the ".is-style-squared" class.