Fix native BlockAlignmentControl (#35191)
* Fix native BlockAlignmentControl

A recent web-centric change introduced usage of `MenuGroup` and
`MenuItem`, both of which are unsupported in the native editor
currently. This updates the conditional within `BlockAlignmentControl`
to reinstate the previous UI for the native platform.

* Fix nuanced test irregularities

Rename mistakenly named assignments, which were misleading.
Additionally, the reliance upon the default `value="left"` could lead to
a false-positive test. Changing to a none-default of `value="right"`
improves the validity of the test.

* Fix inconsistent test description

Co-authored-by: Carlos Garcia <fluiddot@gmail.com>

Co-authored-by: Carlos Garcia <fluiddot@gmail.com>