Remove 'hideImageCaption' context to limit PR's scope
In order to limit this PR's scope, this commit removes the 'hideImageCaption' context. Instead, the focus will be on changes needed for a third-party block to change the image's aspect ratio. The proposal for contexts to hide certain settings in the image block will be moved to https://github.com/WordPress/gutenberg/pull/35315.