Fix missing border in the quick inserter. (#35307)
* Fix missing border in the quick inserter.

In the quick inserter, when no search results are found, there is some left/right border missing. This is caused by some margin applied to the `block-editor-inserter__no-results` div. While this margin could just be removed, I feel it is easier to just apply `overflow: hidden` to the container element. That way the spacing is preserved and should prevent any future issues.

* Update style.scss

Remove top margin and revert previous change.