Test Setup: Add more-complete mocks of common timer functions
The existing mock of `requestAnimationFrame` omits a parameter when
calling the given callback. This patch adds that missing parameter.

Other code in the project also depends on `requestIdleCallback` but
this wasn't mocked at all, leading to test failures in a project
that integrates Gutenberg. This patch adds a reasonable mock of that
function as well.

These mocks have been rearranged as well so that their names match
the expected names by using the more verbose `= function request...`
syntax instead of directly setting their values to the `setTimeout`
functions.