Update logic for checking top-level and child pages
Moves check for top level and child checks to its own function so it
came be a little more clear what it is doing and why.

When the only_child_pages flag is set, we need to treat direct
child pages to the parent as top level so they show. And treat grand
children pages as child pages so they nest.