Fix overflowing content for Help screen
When navigating to a Help section child screen, the parent screen would
remain visible outside of the container. Hiding the overflow of the
containing element fixes this.

The reason this occurred for the first time now is likely because it is
the first time we have utilized the default navigation transitions for
the platforms, which is provided by React Navigation. This means the
screen animates from right to left on iOS, rather than a cross-fade.