Introduce 'fixedHeight' context constant
Following the discussion here, a new 'fixedHeight' context is added to the Gallery block with this commit: https://github.com/WordPress/gutenberg/pull/35510#issuecomment-942248340

This context will be provided by the core Gallery block and possibly other parent blocks in the future. Its default will always be true. It's currently used to set a fixed height on the native Gallery block.