Prepend the `addedCSS` to prevent double `;`
Existing CSS rules already end with a semicolon, and joining it with the `background-color` rule would result in something like:

```
color:#e511be;;background-color:rgba(0, 0, 0, 0)
```

Which is benign AFAIK, but doesn't look right. Best to prepend the `background-color` rule, since it doesn't include a `;`, but the `join` will take care of adding it when joining it with the other existing rules, if any.