Use VStack for a stacked margin approach and remove ability to set a Modal title
Instead of hacking the visuals for the `Modal` component from `ConfirmDialog`, we actually move the logic to hide the title to the `Modal`. It won't render the header if `title` is falsey + a CSS class applied makes sure the additional margin for the header is reset.

Also simplify the `ConfirmDialog` and do not allow passing an optional `title` as the message passed as the `props.children` should be enough.