Refactor to prevent serialisation of context constants
As per the discussion here, part of the purpose of using context constant is to negate the need for serialisation, as the values won't change: https://github.com/WordPress/gutenberg/pull/35510#discussion_r732352036

As such, with this commit, the function that was evaluating the value of the context constants based on a block's attributes is refactored.

With this change, 'providesContextConstants' are added to the block's context object in its entirety, with no need to grab anything from the block's attributes.