Let `wp-scripts packages-update` install packages for a specific version of WordPress.
This is to mitigate the problems related to DEWP usage mentioned in https://github.com/WordPress/gutenberg/issues/35630.

This allows a plugin developer to install to-be-extracted dependencies at the lowest version of their plugin targets. To be able to run local tests and linters against that version. To give some insight to the developer, on what is the lowest anticipated version of an individual package.