[data]: Rename types file from .d.ts to .ts
While working on #35869 I discovered that the module where we keep some
top-level type definitions is named with the `.d.ts` suffix as if it
were a manually-maintained types definition file.

It's capable of being a normal TypeScript module though, and my IDE was
having a bit of trouble with the `.d.ts` suffix - not because that was
wrong, but I think because it was different than what the IDE expected.

In this patch all we're doing is renaming the file in order to smooth
this over and clean up later type-related patches that will occur to
the data package. Updating `tsconfig.json` was necessary becuase of the
rename so that TypeScript will appropriately find and use these types.