Data: Update types and fix type error
In this patch we're making a few updates to the primary types in the data module.
The most important update is a fix. Previously the store config indicated that
the actions and selectors should be an array of functions (or generators) but
those should have been records or objects whose keys were the function names
and whose values where the actual functions. These are equivalent to the type
of the imported modules in practice.

```ts
import * as selectors from './selectors';

type ConfigSelectors = typeof selectors;
```

The names for the core types have been altered to try and simplify their names
and better reflect their usage in the data system.