Revert "Block Library: Fix incorrect attributes definitions #36140"
The `templateLock` attribute is mixed between boolean and string. Setting the type as string introduces an regression that will remove `false` on save.

The original PR (#36140) tried to fix a validation issue reported in #35902

The schema error is incorrect. The attribute definition is required to have either a `type` or an `enum`.