[WIP] Move logic for hiding block controls
With this commit, the logic for hiding block controls is moved to its own function in the 'use-display-block-controls/index.native.' file.

For the purpose of this commit, the code is copied to a native-specific file from the web-version of the file here: https://github.com/WordPress/gutenberg/blob/trunk/packages/block-editor/src/components/use-display-block-controls/index.js

The file will be further refined with native-specific functionality in later commits.