TT | 4058 | Tested scenarios for when the Clipboard Link Suggestion is hidden or pressed in the Link Picker.
0. Hiding the Clipboard Link Suggestion.

Previously, the Clipboard Link Suggestion was only being hidden in the Link Picker if the text in the Clipboard was NOT a valid URL.
Now, the Clipboard Link Suggestion will also be hidden in the Link Picker if the text is a valid URL, but is the same URL the user already picked.

1. Pressing the Clipboard Link Suggestion.

Previously, this was unimplemented.
Now, the behavior is identical to selecting a Link Suggestion from the search results in the Link Picker, i.e., it will navigate back to the previous screen and pass it the selected URL to change the Block Attributes.