Block Styles: Show style preview when hovered or focused  (#34522)
* Initial commit

* Tweaking text and hover colours
Removing commented out code
Consolidating handlers

* Adding container ref to try to position the popover

* Experimenting with removing the default style picker select dropdown with a button click, based on the active style
Updating block supports readme to remove references to the select dropdown
Refactoring DefaultStylePicker
Replacing the transform menu style switcher with a menu

* Abstract renderedStyles, which adds a default to the array for user selection on the frontend.

* Reduce the switcher to a Make default button where the preferred default is not selected.
Added tests for getRenderedStyles

* Aligning preview with preview block popover styles
Creating custom DOMRect object to pass to the PopOver component to set x value.

* Fixed typo in expected label in the getRenderedStyles test

* Cleaning up styles and classnames
`.block-editor-block-styles__item` should exist for the e2e test: style-variation.test.js

* Removing unused preview block code in the block styles menu switcher

* This commit reverts the default style picker, and creates a condition to only show the default picker dropdown where a non-default style has been previously selected. So effectively, we're removing the default style picker from the view. The default style picker will some day be moved to global styles.

* This commit:
- debounces the setHoveredStyle state to avoid flickering effect in the preview when unsetting/setting value
- we need to give more specificity to the classes that hide the preview on narrow widths to avoid overriding
- giving the style buttons fixed widths with ellipses on overflowing text

* Removing `useCallback` function memo

* Using the Text component to control character overflow.
Removing ellipsis styles.

This commit:
- uses InserterPreviewPanel to show the block styles previews, adjusts the attributes being passed to this component so that it displays the correct title.
- remove unused styles

* This commit splits BlockStyles component into two components: one for the settings sidebar and one for inclusion within menu groups.
We are abstracting much of the shared logic to a hook. We've replaced current usages of BlockStyles (2) with the new, specific components.
Also, instead of using `getAnchorRect` to calculate the popover left position, let's overwrite with !important the position of the popover container using the width of the sidebar plus gutter.

* This commit fixes the position of the block preview container and removes the popover.

Adding panel header height to the top position of the preview panel, since there's going to be at least one.

* This is an attempt to ensure that we should the preview panel outside the confines of the sidebar container.
Removing overflow: auto allows us to position elements beyond the container.
We have to detect fullscreen vs non-fullscreen mode to allow for varying top positions.
Also updated selected active states of the button.

* This commit adds a slot in the editor content area for the block styles popover.
We're also reverting changes from the previous commit, where we removed the overflow scroll on the sidebar.
In order to position the preview correctly we need to keep an eye on the scroll top value of the content area.

* Fixing z-index to have the same z-index as the sidebar.
Using useLayoutEffect for setting the scrollTop.

* Using lodash debounce so we can cancel the debounce when needed.
Also hiding preview when mousing out of the preview container.