Rename new support and variable for clarity
The '__experimentalDisplayChildBlockControls' is renamed to '__experimentalHideChildBlockControls' with this commit.

This was suggested here: https://github.com/WordPress/gutenberg/pull/36371#discussion_r747724852

With the following reason as to why this will improve clarity: "I feel it's a bit odd introducing a new block supports that declares a feature that already exists: child block controls. It might be confusing to see blocks that don't include this particular block supports at all, but still display their child's block controls."