Detect if a parent block is hiding child block controls
If a parent block is hiding child block controls, we can reasonable deduce that that would include the option to edit the child block's caption.

This commit builds on the work done in https://github.com/WordPress/gutenberg/pull/36371 to detect if a parent block is hiding child block controls via the '__experimentalHideChildBlockControls' block supports.