[RNMobile] Enable child block caption field to be hidden by parent blocks  (#36618)
This commit provides a way for parent blocks to hide the block caption component (if used) in any inner child block.

The code builds on the work done in #36371, which introduced a new "__experimentalHideChildBlockControls" block support. The purpose of the new block support is to hide child block settings/controls, which is helpful in cases when native blocks need to customize inner blocks. A further outline of the use case for this support can be found here.

With this PR, the block caption component now also listens for "__experimentalHideChildBlockControls". If it is set to true then any captions are hidden from child blocks.