Avoid state updates on unmounted LinkSettingsScreen component
State updates that triggerd from calling `submit` within
`onHandleClosingBottomSheet` resulted in errors from React regarding a
potential memory leak. Since the bottom sheet is closing and component
is unmounting, there is no need to update the component state.

The `onChange` callback, however, still needs to be invoked to persist
the attribute changes. There is further opportunity to improve
`onChange` as it currently empties the attributes causing a re-render
during the close animation, which results in UI jumping.

```
 ERROR  Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.
    in LinkSettingsScreen (at modal.native.js:25)
```