Avoid unnecessary link removal while closing bottom sheet
The absence of a checking the visibility of the link settings meant
multiple calls to `removeLink` were performed while closing the bottom
sheet. This resulted in unnecessary re-renders while the bottom sheet
close animation was running, which led to poor animation performance and
potential for unconsumed `LayoutAnimation`s.

This change now only removes the link if the bottom sheet is still
visible, which matches the behavior of submitting a link.

Unconsumed `LayoutAnimation`s appear to cause issues for React Native's
Modal component on iOS. Specifically, it can result in a transparent,
non-dismissible modal sitting atop the rest of the app UI. Upgrading to
React Native 0.66 appears to have increased the frequency of this
occurring, for an unknown reason.

- https://git.io/J1W3z
- https://git.io/J1W32
- https://git.io/J1W3a
- https://git.io/J1W3r
- https://git.io/J1W36