Tools Panel: Fix race conditions caused by conditionally displayed ToolsPanelItems (#36588)
* Ensure state changes happen in order and flagged values are not overwritten

It's possible to encounter race conditions when updating the panelItems and menuItems
if you have a ToolsPanelItem that's displayed conditionally, or which is shown by
default only conditionally. This can cause multiple calls to update the state in quick
sequence, which can result in those calls overwriting one another. To prevent this,
the calls to setState have been updated to depend on the previous state.

* Add storybook examples

* Update storybook example name

* Do not mutate existing state when registering PanelItems

* Prevent deregistration of panelItems when switching between panels

* Linting

Co-authored-by: Andrew Serong <14988353+andrewserong@users.noreply.github.com>

* Add unit tests for conditionally displayed panel items

* Add unit test for regression with deregistration of panelItems

* Prevent triggering onDeselect

* Clean up tests

* Changelog

* Clarify comments in tests

Co-authored-by: Andrew Serong <14988353+andrewserong@users.noreply.github.com>
Co-authored-by: Aaron Robertshaw <60436221+aaronrobertshaw@users.noreply.github.com>