Fix pagination on empty query
This commit adjusts the logic that handles the display of the `Next Page` link in the Query Loop Pagination block.

The change makes sure that when there are no query results, the `Next Page` link won't be displayed.

This addresses an issue reported at https://github.com/Automattic/wp-calypso/issues/56504