fix scrollIntoView import issue and fix component test case
There was an error saying scrollIntoView function not found. I was not
importing it properly now in this commit I have renamed scrollIntoView function
from dom-scroll-into-view package is renamed as scrollView. I had to do this because
there is already props with the same name ie scrollIntoView. Component test case was
failing because of issue on the same lines that is now fixed.

There was another error related to not using act from react-dom/test-utils package in
test case. I have fixed it. All component test case are passing now.