Remove the semicolons from the JSON examples
I was working through the "Supports" documentation and realised there are some semicolons in the JSON examples. It's best if we remove these so if people are copy and pasting anything into their JSON files it'll work as expected and they don't need to remove the semicolons manually to get webpack to build assets successfully.