Avoid non-dismissible modals caused by unnecessary height transitions
Transitioning height between sub-pixel values is unnecessary and
caused unconsumed LayoutAnimations which resulted in non-dismissible,
transparent modals on iOS. These changes refactor animations for the
`BottomSheetNavigationContainer` to reduce the likelihood of
`LayoutAnimations` remaining unconsumed.

In certain scenarios, the BottomSheet would erroneously register a
`LayoutAnimation`, but then it would never be consumed for one reason
or another. E.g. transitioning to height that is only a few decimal
different.

Unconsumed `LayoutAnimations` appear to cause issues for React Native's
Modal component on iOS. Specifically, it can result in a transparent,
non-dismissible modal sitting atop the rest of the app UI. Upgrading to
React Native 0.66 appears to have increased the frequency of this
occurring, for an unknown reason.

- https://git.io/J1W3z
- https://git.io/J1W32
- https://git.io/J1W3a
- https://git.io/J1W3r
- https://git.io/J1W36