Adjust bottom sheet max height to account for header subtraction
The percentage height required changing now that we subtract the header
height from the max height calculation. The header is now subtracted to
ensure that the `LayoutAnimation` scheduled is consumed if _only_ the
header changes its height. Previously, a change in header height only
scheduled a `LayoutAnimation`, but did not modify the max height.