Update Block Editor Handbook front page
In prep for WP 5.9 update the handbook first page with a less block
focused theme, the original goal for the handbook was to get people in
and creating blocks. There is now so much more from block themes to
patterns, to theme.json that we need to update the content to take all
into account.