Fix BottomSheet navigation hardware back button support
An earlier addition of `setHeight` to the dependencies array meant the
callback was recreated multiple times, which led to unintentional updates
to the hardware button handler. This resulted in navigating backwards
multiple screen instead of one when pressing the hardward back button in
a deeply nested screen.

`onHandleHardwareButtonPress` stores a single value, which means
future invocations from sibling screens can replace the callback for
the currently active screen. Currently, the empty dependency array
passed to `useCallback` here is what prevents erroneous callback
replacements, but leveraging memoization to achieve this is brittle and
explicitly discouraged in the React documentation.
https://reactjs.org/docs/hooks-reference.html#usememo

Ideally, we refactor `onHandleHardwareButtonPress` to manage multiple
callbacks triggered based upon which screen is currently active.

Related: https://git.io/JD2no