Refactor SourceFile extension
With this commit, the SourceFile extension is moved to 'SourceFile.swift' and refactored so that JavaScript is dynamically generated based on file type (e.g. CSS, JS, etc). This refactoring also means that we're now able to revert the change that was made to the 'toJsScript' function's access modifier.

This follows the suggestion here, which also further details the benefits of this approach: https://github.com/WordPress/gutenberg/pull/37221#discussion_r771093813