Update slug for plugin-sidebar tutorial
An idiosyncracy of WP handbook publishing, there can't be a slug conflict,
the changes in https://github.com/WordPress/gutenberg/pull/37490 used
the same slug due to the filename, which should of matched to the
directory.

This conflicted with the existing reference document here:
https://developer.wordpress.org/block-editor/reference-guides/slotfills/plugin-sidebar/

The published URL should actually be: /how-to-guides/plugin-sidebar-0/
which mapped to the old file instead of the directory.

You can confirm by not seeing the Plugin Sidebar in the navigation.