Styling changes; warning icon and colour change
This commit includes changes to the styling of the contrast checker message. As part of the changes, a warning icon is added and the colour of the icon/text is updated to yellow.

As part of these changes, the class names of each element has been updated to match with the BEM methodology, as per core standards.