Add npm-run-all package to simplify running scripts (#37558)
* Add npm-run-all package to simplify running scripts

The npm-run-all package allows for an easier way to run multiple scripts
from the package.json file. We are going to run a few scripts to
generate api docs, so instead of joining multiple commands using && it
can be simplified to creating a command for each and then running:

`npm-run-all api-docs:*`

Package docs:
https://www.npmjs.com/package/npm-run-all

* Update lint-staged (pre-commit hook) with api-docs

Use the npm-run-all in the lint-staged command that gets run by husky on
pre-commit hooks. This allows for checking if any block.json and core
blocks are modified that triggers an update to the reference doc.

* Update package.json

Co-authored-by: Greg Ziółkowski <grzegorz@gziolo.pl>

Co-authored-by: Greg Ziółkowski <grzegorz@gziolo.pl>