Refactor SuggestionsList to use hooks and change class to function component (#36924)
* Refactor SuggestionsList to use hooks and change class to function component

Introduced functional component. In the component I'm using useState, useEffect.
Renamed a few variables to get rid of syntax errors. Tested the component and
it is working as expected.

* fix scrollIntoView import issue and fix component test case

There was an error saying scrollIntoView function not found. I was not
importing it properly now in this commit I have renamed scrollIntoView function
from dom-scroll-into-view package is renamed as scrollView. I had to do this because
there is already props with the same name ie scrollIntoView. Component test case was
failing because of issue on the same lines that is now fixed.

There was another error related to not using act from react-dom/test-utils package in
test case. I have fixed it. All component test case are passing now.

* Address review comments

- Add `noop` to reduce number of functions creation
- Move  emptyList to a const then use it
- Move `handleMouseDown` to top of the file to avoid recreating it on every render

* refactor `noop`, set default value `scrollingIntoView` false and use `useRefEffect`

* add changelog entry

* set `onHover` and `onSelect` default value undefined and use optional chaining