Block API: Separate validation stage during block parsing
When initially parsing and loading blocks we attempt to validate their parsed
attributes and apply auto-fixes when invalid. Previously this has taken place
inline inside of `parseRawBlock()`.

In this patch we're separating out the validation stage to aid in understanding
what actions take place during the validation and to allow `parseRawBlock()` to
avoid swapping the value of `parsedBlock` while running. With this patch each
separate stage in the parse is stored in a separate `const` variable which
enables easier debugging and comparison.

Additionaly the condition for printing warnings on block invalidation have been
updated to more clearly reflect the logic triggering them: that there were
problems during validation that were or were not resolved through the
block deprecation updating stage.