Remove logic to discard unsaved changes when last Site Logo removed
This effect ran on unmount and used `getGlobalBlockCount` to determine
if all Site Logo blocks have been removed, then discards any unsaved
changes to the Logo and Icon.

This does not work correctly in the Site Editor context, and would fire
erroneously when switching between template parts, breaking the ability
to 'Undo'/'Redo'.