docgen: Stop crashing on missing return types
In certain situations if `docgen` finds a JSDoc with a description
for the return value of a function but there's no corresponding
type listed for the return value it will error out and say it's required.
This requirement introduces a mismatch with how many TypeScript functions
will be idiomatically _typed_. In this patch we're reporting those missing
types as _unknown_ and letting the `docgen` continue.

When missing return types of default exports `docgen` would silently
crash without an error message because of a bug when trying to find the
name of the surrounding function _when generating the error message_.
This is fixed by special-casing default exports alongside named exports.