Wrap long cell labels on Android
Prior to this change, the cell label container was disallowed from
shrinking to accommodate its siblings. This caused longer labels to
force the sibling elements out of the viewport causing them to be
invisible or cut-off.

Additional changes were applied to the link suggestion styles to ensure
the sibling of cell icons do not encroach upon the now static width
icon.