Wrap long cell labels on Android (#37993)
* Wrap long cell labels on Android

Prior to this change, the cell label container was disallowed from
shrinking to accommodate its siblings. This caused longer labels to
force the sibling elements out of the viewport causing them to be
invisible or cut-off.

Additional changes were applied to the link suggestion styles to ensure
the sibling of cell icons do not encroach upon the now static width
icon.

* Add change log entry

* Remove unnecessary style declarations

These styles are the default values and redundant.

* Expand link suggestion title to parent width

Force link suggestion title to fill parent width, which causes the link
suggestion to match the parent separator line width.