[Docs]: Add docs for schema in block API
While the `schema` property is mentioned in the block API documentation
for `raw` transforms it's not explained.

In this patch we're adding a short sub-section to expand on the purpose
and implementation of the `schema` in a way that can hopefully give
developers a stronger starting point for when they need to implement
new raw transformations.

Co-authored-by: sarayourfriend <24264157+sarayourfriend@users.noreply.github.com>