[docgen] Refactor code to use require() instead of JSON.parse(fs.readFileSync())
While working through #37929 I noticed that the tests in the `docgen` package
read fixture data manually by chaining `JSON.parse()` and `fs.readFileSync()`,
often at times separated by a number of lines.

In this patch we're replacing those with the more natural `require()` statement
which removes noise from the test modules and hopefully will also remove a point
of confusion for those coming in to modify it. We shouldn't have to ask, "why is
this code doing something that looks normal but is doing it in a notably different
manner?"

After reviewing the history of the work, initially introduced in #13329, I could
find no explanation for the approach and found no discussion about it in the PR.